requirements.txt
# TradeWise MVP - Dependencies
# Python 3.10+ required

streamlit>=1.28.0
yfinance>=0.2.36
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
